requirements.txt
torch
torchvision
Pillow
gradio
torchtext
opencv-python
moviepy